hot topic pagenation
